requirements.txt
flask
flask-cors
requests
PyPDF2
python-docx
